move branch drop to CutOffConcerns, add handling for empty fringe.
