+++ org.palladiosimulator.analyzer.slingshot.stateexploration/src/org/palladiosimulator/analyzer/slingshot/stateexploration/explorer/planning/CutOffConcerns.java
@@ -1,7 +1,10 @@
 package org.palladiosimulator.analyzer.slingshot.stateexploration.explorer.planning;
 
 import org.apache.log4j.Logger;
+import org.palladiosimulator.analyzer.slingshot.stateexploration.api.RawTransition;
+import org.palladiosimulator.analyzer.slingshot.stateexploration.rawgraph.DefaultState;
 import org.palladiosimulator.analyzer.slingshot.stateexploration.rawgraph.ToDoChange;
+import org.palladiosimulator.spd.ScalingPolicy;
 
 /**
  * TODO
@@ -19,6 +22,54 @@ public class CutOffConcerns {
 		LOGGER.debug(String.format("Evaluation future %s.", future.toString()));
 		LOGGER.debug(String.format("Future %s is rosy, will explore.", future.toString()));
 
-		return true;
+		return matchesPattern(future);
+	}
+
+	/**
+	 * Patter is "leav on rea" (prev) -> NOOP -> "leav on rea" (current) -> NOOP
+	 * (ToDoChange)
+	 *
+	 * @param current
+	 * @return
+	 */
+	private boolean matchesPattern(final ToDoChange change) {
+
+		final DefaultState current = change.getStart();
+
+		if (current.getIncomingTransition().isEmpty()) { // root?
+			return false;
+		}
+
+		final DefaultState prev = (DefaultState) current.getIncomingTransition().get().getSource();
+
+		return samePolicy(current, prev) && bothNOOP(change, current.getIncomingTransition().get());
+	}
+
+	/**
+	 *
+	 * @param current
+	 * @param prev
+	 * @return
+	 */
+	private static boolean bothNOOP(final ToDoChange current, final RawTransition prev) {
+		return current.getChange().isEmpty() && prev.getChange().isEmpty();
+	}
+
+	/**
+	 *
+	 * @param current
+	 * @param prev
+	 * @return
+	 */
+	private static boolean samePolicy(final DefaultState current, final DefaultState prev) {
+		if (current.getSnapshot().getModelAdjustmentRequestedEvent().isEmpty()
+				|| prev.getSnapshot().getModelAdjustmentRequestedEvent().isEmpty()) {
+			return false;
+		}
+		final ScalingPolicy policyCurrent = current.getSnapshot().getModelAdjustmentRequestedEvent().get()
+				.getScalingPolicy();
+		final ScalingPolicy policyPrev = prev.getSnapshot().getModelAdjustmentRequestedEvent().get().getScalingPolicy();
+
+		return policyCurrent.getId().equals(policyPrev.getId());
 	}
 }
